Fixes lost resource production when opening a popup When a player views a playercard or information popup of building etc, it should be countet as action resulting in an updated * in galaxy map. This was acheved by updating the planets timestamp directly. This unfortunately means: Wait 5 min on a page, open a popup and you lost 5 min of ress production.
With this change the timestamp will be updated only by calculating the resources. Most popups will trigger the calculation.
Error pages and battle reports will no longer update the timestamp but will also not calculate the resources and therefore will not be visible as user action to other users via galaxy.